LayoutTests/fast/js/exception-linenums-in-html-1-expected.txt

This test exercises the source URL and line number that is embedded in JavaScript exceptions, which is displayed in places like the JavaScript console. It differs from exception-linenums.html in that the line numbers are derived from within the HTML source, not a JavaScript source file.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof e.sourceURL is "string"
PASS e.line is 19
PASS typeof e.sourceURL is "string"
PASS e.line is 29
PASS typeof e.sourceURL is "string"
PASS e.line is 39
PASS typeof e.sourceURL is "string"
PASS e.line is 49
PASS typeof e.sourceURL is "string"
PASS e.line is 12
PASS successfullyParsed is true

TEST COMPLETE
